pyproject.toml
[project]
name = "memory-local"
version = "0.1.0"
description = "Self-hosted memory system using Mem0, PostgreSQL/pgvector, and Ollama"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-neo4j>=0.9.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
all = [
    "mem0ai",
    "ollama",
    "psycopg2-binary",
    "requests",
    "neo4j",
    "python-dotenv",
    "qdrant-client",
]

[project.scripts]
memory-local = "memory_local.setup:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memory_local"]
